[buildscript] force chosen phase to be executed (GNOME bug 583515)
If there is an error and a new phase is selected (be it by the user or
an automatic system), the chosen phase must absolutely be executed, it
should in no condition be skipped automatically.  This commit introduces
a new force_phase variable that identifies this situation and skip the
skip check.